requirements.txt
numpy
scipy
pip
fake-bpy-module